+++ RELEASENOTES.md
@@ -31,6 +31,18 @@ These are changes that have backwards-compatible solutions present and that comp
 
 ### Latest Release
 
+## Release 0.10.16
+
+**Minor Features**
+
+- `Datepicker`: Add `canInputMaintainFocus` prop
+- `Tooltip`: Add `onRequestTargetElement` callback
+- `Tooltip`: Add `hasAnchoredNubbin` prop
+
+**Bugfixes**
+
+- `Colorpicker`: Add support for named colors in default validator
+
 ## Release 0.10.15
 
 **Bugfixes**